# docs/swagger.yaml
basePath: /api/v1/device
definitions:
  types.CreateDevicePayload:
    properties:
      brand:
        type: string
      id:
        type: integer
      name:
        type: string
      state:
        type: string
    required:
    - brand
    - id
    - name
    - state
    type: object
  types.Device:
    properties:
      brand:
        type: string
      creation_time:
        type: string
      id:
        type: integer
      name:
        type: string
      state:
        type: string
    type: object
  types.UpdateDevicePayload:
    properties:
      brand:
        type: string
      id:
        type: integer
      name:
        type: string
      state:
        type: string
    required:
    - id
    type: object
host: localhost:8080
info:
  contact: {}
  description: API used for device management in this code challenge
  title: Device API
  version: "1.0"
paths:
  /createDevice:
    post:
      consumes:
      - application/json
      description: Create a new device
      parameters:
      - description: Device Id
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/types.CreateDevicePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.Device'
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      summary: Create a new device
      tags:
      - Devices
  /deleteDevice/{id}:
    patch:
      description: Delete the provided device
      parameters:
      - description: device ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: delete the device
      tags:
      - Devices
  /getDeviceById/{Id}:
    get:
      description: Return a device by its id
      parameters:
      - description: device ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Device'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Search for a specific device
      tags:
      - Devices
  /getDevices:
    get:
      description: Return a list of all devices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Device'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Search for all devices
      tags:
      - Devices
  /getDevicesByBrand/{brand}:
    get:
      description: Return a list of all devices with the specific brand
      parameters:
      - description: device brand
        in: path
        name: brand
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Device'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Search for all devices with the specific brand
      tags:
      - Devices
  /getDevicesByState/{state}:
    get:
      description: Return a list of all devices with the specific state
      parameters:
      - description: device state
        in: path
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Device'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Search for all devices with the specific state
      tags:
      - Devices
  /updateDevice:
    patch:
      consumes:
      - application/json
      description: Udate only provided fields and with no "in-use" state
      parameters:
      - description: Device Id
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/types.UpdateDevicePayload'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: update the device
      tags:
      - Devices
swagger: "2.0"